refactor: extract VoucherWalletOverlay from Dashboard
